+++ imgaug/augmenters/convolutional.py
@@ -1,5 +1,5 @@
 """
-Augmenters that apply convolutions to images.
+Augmenters that are based on applying convolution kernels to images.
 
 Do not import directly from this file, as the categorization is not final.
 Use instead ::
@@ -27,6 +27,7 @@ For MotionBlur, see ``blur.py``.
 from __future__ import print_function, division, absolute_import
 
 import types
+import itertools
 
 import numpy as np
 import cv2
@@ -38,9 +39,10 @@ from .. import parameters as iap
 from .. import dtypes as iadt
 
 
+# TODO allow 3d matrices as input (not only 2D)
 class Convolve(meta.Augmenter):
     """
-    Apply a Convolution to input images.
+    Apply a convolution to input images.
 
     dtype support::
 
@@ -59,9 +61,10 @@ class Convolve(meta.Augmenter):
         * ``bool``: yes; tested (4)
 
         - (1) rejected by ``cv2.filter2D()``.
-        - (2) causes error: cv2.error: OpenCV(3.4.2) (...)/filter.cpp:4487: error: (-213:The function/feature is
-              not implemented) Unsupported combination of source format (=1), and destination format (=1) in
-              function 'getLinearFilter'.
+        - (2) causes error: cv2.error: OpenCV(3.4.2) (...)/filter.cpp:4487:
+              error: (-213:The function/feature is not implemented)
+              Unsupported combination of source format (=1), and destination
+              format (=1) in function 'getLinearFilter'.
         - (3) mapped internally to ``int16``.
         - (4) mapped internally to ``float32``.
 
@@ -70,15 +73,16 @@ class Convolve(meta.Augmenter):
     matrix : None or (H, W) ndarray or imgaug.parameters.StochasticParameter or callable, optional
         The weight matrix of the convolution kernel to apply.
 
-            * If None, the input images will not be changed.
-            * If a numpy array, that array will be used for all images and
-              channels as the kernel.
-            * If a callable, the parameter will be called for each image
-              via ``param(image, C, random_state)``. The function must either return
-              a list of ``C`` matrices (i.e. one per channel) or a 2D numpy array
-              (will be used for all channels) or a 3D ``HxWxC`` numpy array.
-              If a list is returned, each entry may be None, which will result
-              in no changes to the respective channel.
+            * If ``None``, the input images will not be changed.
+            * If a 2D numpy array, that array will always be used for all
+              images and channels as the kernel.
+            * If a callable, that method will be called for each image
+              via ``parameter(image, C, random_state)``. The function must
+              either return a list of ``C`` matrices (i.e. one per channel)
+              or a 2D numpy array (will be used for all channels) or a
+              3D ``HxWxC`` numpy array. If a list is returned, each entry may
+              be ``None``, which will result in no changes to the respective
+              channel.
 
     name : None or str, optional
         See :func:`imgaug.augmenters.meta.Augmenter.__init__`.
@@ -91,12 +95,13 @@ class Convolve(meta.Augmenter):
 
     Examples
     --------
+    >>> import imgaug.augmenters as iaa
     >>> matrix = np.array([[0, -1, 0],
     >>>                    [-1, 4, -1],
     >>>                    [0, -1, 0]])
     >>> aug = iaa.Convolve(matrix=matrix)
 
-    convolves all input images with the kernel shown in the `matrix`
+    Convolves all input images with the kernel shown in the ``matrix``
     variable.
 
     >>> def gen_matrix(image, nb_channels, random_state):
@@ -112,39 +117,48 @@ class Convolve(meta.Augmenter):
     >>>         return [matrix_B] * nb_channels
     >>> aug = iaa.Convolve(matrix=gen_matrix)
 
-    convolves images that have an even height with matrix A and images
-    with an odd height with matrix B.
+    Convolves images that have an even height with matrix A and images
+    having an odd height with matrix B.
 
     """
 
-    def __init__(self, matrix=None, name=None, deterministic=False, random_state=None):
-        super(Convolve, self).__init__(name=name, deterministic=deterministic, random_state=random_state)
+    def __init__(self, matrix=None,
+                 name=None, deterministic=False, random_state=None):
+        super(Convolve, self).__init__(
+            name=name, deterministic=deterministic, random_state=random_state)
 
         if matrix is None:
             self.matrix = None
             self.matrix_type = "None"
         elif ia.is_np_array(matrix):
-            ia.do_assert(len(matrix.shape) == 2,
-                         "Expected convolution matrix to have 2 axis, got %d (shape %s)." % (
-                             len(matrix.shape), matrix.shape))
+            ia.do_assert(
+                matrix.ndim == 2,
+                "Expected convolution matrix to have exactly two dimensions, "
+                "got %d (shape %s)." % (
+                    matrix.ndim, matrix.shape))
             self.matrix = matrix
             self.matrix_type = "constant"
         elif isinstance(matrix, types.FunctionType):
             self.matrix = matrix
             self.matrix_type = "function"
         else:
-            raise Exception("Expected float, int, tuple/list with 2 entries or StochasticParameter. Got %s." % (
-                type(matrix),))
+            raise Exception(
+                "Expected float, int, tuple/list with 2 entries or "
+                "StochasticParameter. Got %s." % (
+                    type(matrix),))
 
     def _augment_images(self, images, random_state, parents, hooks):
         iadt.gate_dtypes(images,
-                         allowed=["bool", "uint8", "uint16", "int8", "int16", "float16", "float32", "float64"],
+                         allowed=["bool",
+                                  "uint8", "uint16",
+                                  "int8", "int16",
+                                  "float16", "float32", "float64"],
                          disallowed=["uint32", "uint64", "uint128", "uint256",
                                      "int32", "int64", "int128", "int256",
-                                      "float96", "float128", "float256"],
+                                     "float96", "float128", "float256"],
                          augmenter=self)
+        rss = ia.derive_random_states(random_state, len(images))
 
-        seed = random_state.randint(0, 10**6, 1)[0]
         for i, image in enumerate(images):
             _height, _width, nb_channels = images[i].shape
 
@@ -159,22 +173,32 @@ class Convolve(meta.Augmenter):
             elif self.matrix_type == "constant":
                 matrices = [self.matrix] * nb_channels
             elif self.matrix_type == "function":
-                matrices = self.matrix(images[i], nb_channels, ia.new_random_state(seed+i))
+                matrices = self.matrix(images[i], nb_channels, rss[i])
                 if ia.is_np_array(matrices) and matrices.ndim == 2:
-                    matrices = np.tile(matrices[..., np.newaxis], (1, 1, nb_channels))
+                    matrices = np.tile(
+                        matrices[..., np.newaxis],
+                        (1, 1, nb_channels))
+
+                is_valid_list = (isinstance(matrices, list)
+                                 and len(matrices) == nb_channels)
+                is_valid_array = (ia.is_np_array(matrices)
+                                  and matrices.ndim == 3
+                                  and matrices.shape[2] == nb_channels)
                 ia.do_assert(
-                    (isinstance(matrices, list) and len(matrices) == nb_channels)
-                    or (ia.is_np_array(matrices) and matrices.ndim == 3 and matrices.shape[2] == nb_channels),
-                    "Callable provided to Convole must return either a list of 2D matrices (one per image channel) "
+                    is_valid_list or is_valid_array,
+                    "Callable provided to Convole must return either a "
+                    "list of 2D matrices (one per image channel) "
                     "or a 2D numpy array "
-                    "or a 3D numpy array where the last dimension's size matches the number of image channels. "
+                    "or a 3D numpy array where the last dimension's size "
+                    "matches the number of image channels. "
                     "Got type %s." % (type(matrices),)
                 )
 
                 if ia.is_np_array(matrices):
-                    # Shape of matrices is currently (H, W, C), but in the loop below we need the
-                    # first axis to be the channel index to unify handling of lists of arrays
-                    # and arrays. So we move the channel axis here to the start.
+                    # Shape of matrices is currently (H, W, C), but in the
+                    # loop below we need the first axis to be the channel
+                    # index to unify handling of lists of arrays and arrays.
+                    # So we move the channel axis here to the start.
                     matrices = matrices.transpose((2, 0, 1))
             else:
                 raise Exception("Invalid matrix type")
@@ -182,9 +206,13 @@ class Convolve(meta.Augmenter):
             image_aug = image
             for channel in sm.xrange(nb_channels):
                 if matrices[channel] is not None:
-                    # ndimage.convolve caused problems here
-                    # cv2.filter2D() always returns same output dtype as input dtype
-                    image_aug[..., channel] = cv2.filter2D(image_aug[..., channel], -1, matrices[channel])
+                    # ndimage.convolve caused problems here cv2.filter2D()
+                    # always returns same output dtype as input dtype
+                    image_aug[..., channel] = cv2.filter2D(
+                        image_aug[..., channel],
+                        -1,
+                        matrices[channel]
+                    )
 
             if input_dtype == np.bool_:
                 image_aug = image_aug > 0.5
@@ -196,10 +224,13 @@ class Convolve(meta.Augmenter):
         return images
 
     def _augment_heatmaps(self, heatmaps, random_state, parents, hooks):
+        # pylint: disable=no-self-use
         # TODO this can fail for some matrices, e.g. [[0, 0, 1]]
         return heatmaps
 
-    def _augment_keypoints(self, keypoints_on_images, random_state, parents, hooks):
+    def _augment_keypoints(self, keypoints_on_images, random_state, parents,
+                           hooks):
+        # pylint: disable=no-self-use
         # TODO this can fail for some matrices, e.g. [[0, 0, 1]]
         return keypoints_on_images
 
@@ -207,9 +238,10 @@ class Convolve(meta.Augmenter):
         return [self.matrix, self.matrix_type]
 
 
-def Sharpen(alpha=0, lightness=1, name=None, deterministic=False, random_state=None):
+def Sharpen(alpha=0, lightness=1,
+            name=None, deterministic=False, random_state=None):
     """
-    Augmenter that sharpens images and overlays the result with the original image.
+    Sharpen images and alpha-blend the result with the original input images.
 
     dtype support::
 
@@ -218,29 +250,29 @@ def Sharpen(alpha=0, lightness=1, name=None, deterministic=False, random_state=N
     Parameters
     ----------
     alpha : number or tuple of number or list of number or imgaug.parameters.StochasticParameter, optional
-        Visibility of the sharpened image. At 0, only the original image is
-        visible, at 1.0 only its sharpened version is visible.
+        Blending factor of the sharpened image. At ``0.0``, only the original
+        image is visible, at ``1.0`` only its sharpened version is visible.
 
-            * If an int or float, exactly that value will be used.
-            * If a tuple ``(a, b)``, a random value from the range ``a <= x <= b`` will
-              be sampled per image.
-            * If a list, then a random value will be sampled from that list
+            * If a number, exactly that value will always be used.
+            * If a tuple ``(a, b)``, a random value will be sampled from the
+              interval ``[a, b]`` per image.
+            * If a list, a random value will be sampled from that list
               per image.
-            * If a StochasticParameter, a value will be sampled from the
+            * If a ``StochasticParameter``, a value will be sampled from that
               parameter per image.
 
     lightness : number or tuple of number or list of number or imgaug.parameters.StochasticParameter, optional
-        Parameter that controls the lightness/brightness of the sharped image.
-        Sane values are somewhere in the range ``(0.5, 2)``.
-        The value 0 results in an edge map. Values higher than 1 create bright
-        images. Default value is 1.
+        Lightness/brightness of the sharped image.
+        Sane values are somewhere in the interval ``[0.5, 2.0]``.
+        The value ``0.0`` results in an edge map. Values higher than ``1.0``
+        create bright images. Default value is ``1.0``.
 
-            * If an int or float, exactly that value will be used.
-            * If a tuple ``(a, b)``, a random value from the range ``a <= x <= b`` will
-              be sampled per image.
-            * If a list, then a random value will be sampled from that list
+            * If a number, exactly that value will always be used.
+            * If a tuple ``(a, b)``, a random value will be sampled from the
+              interval ``[a, b]`` per image.
+            * If a list, a random value will be sampled from that list
               per image.
-            * If a StochasticParameter, a value will be sampled from the
+            * If a ``StochasticParameter``, a value will be sampled from that
               parameter per image.
 
     name : None or str, optional
@@ -254,26 +286,32 @@ def Sharpen(alpha=0, lightness=1, name=None, deterministic=False, random_state=N
 
     Examples
     --------
-    >>> aug = Sharpen(alpha=(0.0, 1.0))
+    >>> import imgaug.augmenters as iaa
+    >>> aug = iaa.Sharpen(alpha=(0.0, 1.0))
 
-    sharpens input images and overlays the sharpened image by a variable
-    amount over the old image.
+    Sharpens input images and blends the sharpened image with the input image
+    using a random blending factor between ``0%`` and ``100%`` (uniformly
+    sampled).
 
-    >>> aug = Sharpen(alpha=(0.0, 1.0), lightness=(0.75, 2.0))
+    >>> aug = iaa.Sharpen(alpha=(0.0, 1.0), lightness=(0.75, 2.0))
 
-    sharpens input images with a variable lightness in the range
-    ``0.75 <= x <= 2.0`` and with a variable alpha.
+    Sharpens input images with a variable `lightness` sampled uniformly from
+    the interval ``[0.75, 2.0]`` and with a fully random blending factor
+    (as in the above example).
 
     """
-    alpha_param = iap.handle_continuous_param(alpha, "alpha", value_range=(0, 1.0), tuple_to_uniform=True,
-                                              list_to_choice=True)
-    lightness_param = iap.handle_continuous_param(lightness, "lightness", value_range=(0, None), tuple_to_uniform=True,
-                                                  list_to_choice=True)
+    alpha_param = iap.handle_continuous_param(
+        alpha, "alpha",
+        value_range=(0, 1.0), tuple_to_uniform=True, list_to_choice=True)
+    lightness_param = iap.handle_continuous_param(
+        lightness, "lightness",
+        value_range=(0, None), tuple_to_uniform=True, list_to_choice=True)
 
     def create_matrices(image, nb_channels, random_state_func):
         alpha_sample = alpha_param.draw_sample(random_state=random_state_func)
         ia.do_assert(0 <= alpha_sample <= 1.0)
-        lightness_sample = lightness_param.draw_sample(random_state=random_state_func)
+        lightness_sample = lightness_param.draw_sample(
+            random_state=random_state_func)
         matrix_nochange = np.array([
             [0, 0, 0],
             [0, 1, 0],
@@ -284,19 +322,26 @@ def Sharpen(alpha=0, lightness=1, name=None, deterministic=False, random_state=N
             [-1, 8+lightness_sample, -1],
             [-1, -1, -1]
         ], dtype=np.float32)
-        matrix = (1-alpha_sample) * matrix_nochange + alpha_sample * matrix_effect
+        matrix = (
+            (1-alpha_sample) * matrix_nochange
+            + alpha_sample * matrix_effect
+        )
         return [matrix] * nb_channels
 
     if name is None:
         name = "Unnamed%s" % (ia.caller_name(),)
 
-    return Convolve(create_matrices, name=name, deterministic=deterministic, random_state=random_state)
+    return Convolve(
+        create_matrices,
+        name=name,
+        deterministic=deterministic,
+        random_state=random_state)
 
 
-def Emboss(alpha=0, strength=1, name=None, deterministic=False, random_state=None):
+def Emboss(alpha=0, strength=1,
+           name=None, deterministic=False, random_state=None):
     """
-    Augmenter that embosses images and overlays the result with the original
-    image.
+    Emboss images and alpha-blend the result with the original input images.
 
     The embossed version pronounces highlights and shadows,
     letting the image look as if it was recreated on a metal plate ("embossed").
@@ -308,28 +353,28 @@ def Emboss(alpha=0, strength=1, name=None, deterministic=False, random_state=Non
     Parameters
     ----------
     alpha : number or tuple of number or list of number or imgaug.parameters.StochasticParameter, optional
-        Visibility of the sharpened image. At 0, only the original image is
-        visible, at 1.0 only its sharpened version is visible.
+        Blending factor of the embossed image. At ``0.0``, only the original
+        image is visible, at ``1.0`` only its embossed version is visible.
 
-            * If an int or float, exactly that value will be used.
-            * If a tuple ``(a, b)``, a random value from the range ``a <= x <= b`` will
-              be sampled per image.
-            * If a list, then a random value will be sampled from that list
+            * If a number, exactly that value will always be used.
+            * If a tuple ``(a, b)``, a random value will be sampled from the
+              interval ``[a, b]`` per image.
+            * If a list, a random value will be sampled from that list
               per image.
-            * If a StochasticParameter, a value will be sampled from the
+            * If a ``StochasticParameter``, a value will be sampled from that
               parameter per image.
 
     strength : number or tuple of number or list of number or imgaug.parameters.StochasticParameter, optional
         Parameter that controls the strength of the embossing.
-        Sane values are somewhere in the range ``(0, 2)`` with 1 being the standard
-        embossing effect. Default value is 1.
+        Sane values are somewhere in the interval ``[0.0, 2.0]`` with ``1.0``
+        being the standard embossing effect. Default value is ``1.0``.
 
-            * If an int or float, exactly that value will be used.
-            * If a tuple ``(a, b)``, a random value from the range ``a <= x <= b`` will
-              be sampled per image.
+            * If a number, exactly that value will always be used.
+            * If a tuple ``(a, b)``, a random value will be sampled from the
+              interval ``[a, b]`` per image.
             * If a list, then a random value will be sampled from that list
               per image.
-            * If a StochasticParameter, a value will be sampled from the
+            * If a ``StochasticParameter``, a value will be sampled from the
               parameter per image.
 
     name : None or str, optional
@@ -343,22 +388,26 @@ def Emboss(alpha=0, strength=1, name=None, deterministic=False, random_state=Non
 
     Examples
     --------
-    >>> aug = Emboss(alpha=(0.0, 1.0), strength=(0.5, 1.5))
+    >>> import imgaug.augmenters as iaa
+    >>> aug = iaa.Emboss(alpha=(0.0, 1.0), strength=(0.5, 1.5))
 
-    embosses an image with a variable strength in the range ``0.5 <= x <= 1.5``
-    and overlays the result with a variable alpha in the range ``0.0 <= a <= 1.0``
-    over the old image.
+    Emboss an image with a strength sampled uniformly from the interval
+    ``[0.5, 1.5]`` and alpha-blend the result with the original input image
+    using a random blending factor between ``0%`` and ``100%``.
 
     """
-    alpha_param = iap.handle_continuous_param(alpha, "alpha", value_range=(0, 1.0), tuple_to_uniform=True,
-                                              list_to_choice=True)
-    strength_param = iap.handle_continuous_param(strength, "strength", value_range=(0, None), tuple_to_uniform=True,
-                                                 list_to_choice=True)
+    alpha_param = iap.handle_continuous_param(
+        alpha, "alpha",
+        value_range=(0, 1.0), tuple_to_uniform=True, list_to_choice=True)
+    strength_param = iap.handle_continuous_param(
+        strength, "strength",
+        value_range=(0, None), tuple_to_uniform=True, list_to_choice=True)
 
     def create_matrices(image, nb_channels, random_state_func):
         alpha_sample = alpha_param.draw_sample(random_state=random_state_func)
         ia.do_assert(0 <= alpha_sample <= 1.0)
-        strength_sample = strength_param.draw_sample(random_state=random_state_func)
+        strength_sample = strength_param.draw_sample(
+            random_state=random_state_func)
         matrix_nochange = np.array([
             [0, 0, 0],
             [0, 1, 0],
@@ -369,21 +418,27 @@ def Emboss(alpha=0, strength=1, name=None, deterministic=False, random_state=Non
             [0-strength_sample, 1, 0+strength_sample],
             [0, 0+strength_sample, 1+strength_sample]
         ], dtype=np.float32)
-        matrix = (1-alpha_sample) * matrix_nochange + alpha_sample * matrix_effect
+        matrix = (
+            (1-alpha_sample) * matrix_nochange
+            + alpha_sample * matrix_effect
+        )
         return [matrix] * nb_channels
 
     if name is None:
         name = "Unnamed%s" % (ia.caller_name(),)
 
-    return Convolve(create_matrices, name=name, deterministic=deterministic, random_state=random_state)
+    return Convolve(
+        create_matrices,
+        name=name,
+        deterministic=deterministic,
+        random_state=random_state)
 
 
-# TODO tests
+# TODO add tests
+# TODO move this to edges.py?
 def EdgeDetect(alpha=0, name=None, deterministic=False, random_state=None):
     """
-    Augmenter that detects all edges in images, marks them in
-    a black and white image and then overlays the result with the original
-    image.
+    Generate a black & white edge image and alpha-blend it with the input image.
 
     dtype support::
 
@@ -392,15 +447,15 @@ def EdgeDetect(alpha=0, name=None, deterministic=False, random_state=None):
     Parameters
     ----------
     alpha : number or tuple of number or list of number or imgaug.parameters.StochasticParameter, optional
-        Visibility of the sharpened image. At 0, only the original image is
-        visible, at 1.0 only its sharpened version is visible.
+        Blending factor of the edge image. At ``0.0``, only the original
+        image is visible, at ``1.0`` only the edge image is visible.
 
-            * If an int or float, exactly that value will be used.
-            * If a tuple ``(a, b)``, a random value from the range ``a <= x <= b`` will
-              be sampled per image.
-            * If a list, then a random value will be sampled from that list
+            * If a number, exactly that value will always be used.
+            * If a tuple ``(a, b)``, a random value will be sampled from the
+              interval ``[a, b]`` per image.
+            * If a list, a random value will be sampled from that list
               per image.
-            * If a StochasticParameter, a value will be sampled from the
+            * If a ``StochasticParameter``, a value will be sampled from that
               parameter per image.
 
     name : None or str, optional
@@ -414,14 +469,17 @@ def EdgeDetect(alpha=0, name=None, deterministic=False, random_state=None):
 
     Examples
     --------
-    >>> aug = EdgeDetect(alpha=(0.0, 1.0))
+    >>> import imgaug.augmenters as iaa
+    >>> aug = iaa.EdgeDetect(alpha=(0.0, 1.0))
 
-    detects edges in an image  and overlays the result with a variable alpha
-    in the range ``0.0 <= a <= 1.0`` over the old image.
+    Detect edges in an image, mark them as black (non-edge) and white (edges)
+    and alpha-blend the result with the original input image using a random
+    blending factor between ``0%`` and ``100%``.
 
     """
-    alpha_param = iap.handle_continuous_param(alpha, "alpha", value_range=(0, 1.0), tuple_to_uniform=True,
-                                              list_to_choice=True)
+    alpha_param = iap.handle_continuous_param(
+        alpha, "alpha",
+        value_range=(0, 1.0), tuple_to_uniform=True, list_to_choice=True)
 
     def create_matrices(_image, nb_channels, random_state_func):
         alpha_sample = alpha_param.draw_sample(random_state=random_state_func)
@@ -436,22 +494,38 @@ def EdgeDetect(alpha=0, name=None, deterministic=False, random_state=None):
             [1, -4, 1],
             [0, 1, 0]
         ], dtype=np.float32)
-        matrix = (1-alpha_sample) * matrix_nochange + alpha_sample * matrix_effect
+        matrix = (
+            (1-alpha_sample) * matrix_nochange
+            + alpha_sample * matrix_effect
+        )
         return [matrix] * nb_channels
 
     if name is None:
         name = "Unnamed%s" % (ia.caller_name(),)
 
-    return Convolve(create_matrices, name=name, deterministic=deterministic, random_state=random_state)
+    return Convolve(
+        create_matrices,
+        name=name,
+        deterministic=deterministic,
+        random_state=random_state)
 
 
-# TODO tests
+# TODO add tests
 # TODO merge EdgeDetect and DirectedEdgeDetect?
-def DirectedEdgeDetect(alpha=0, direction=(0.0, 1.0), name=None, deterministic=False, random_state=None):
+# TODO deprecate and rename to AngledEdgeDetect
+# TODO rename arg "direction" to "angle"
+# TODO change direction/angle value range to (0, 360)
+# TODO move this to edges.py?
+def DirectedEdgeDetect(alpha=0, direction=(0.0, 1.0),
+                       name=None, deterministic=False, random_state=None):
     """
-    Augmenter that detects edges that have certain directions and marks them
-    in a black and white image and then overlays the result with the original
-    image.
+    Detect edges from specified angles and alpha-blend with the input image.
+
+    This augmenter first detects edges along a certain angle.
+    Usually, edges are detected in x- or y-direction, while here the edge
+    detection kernel is rotated to match a specified angle.
+    The result of applying the kernel is a black (non-edges) and white (edges)
+    image. That image is alpha-blended with the input image.
 
     dtype support::
 
@@ -460,28 +534,29 @@ def DirectedEdgeDetect(alpha=0, direction=(0.0, 1.0), name=None, deterministic=F
     Parameters
     ----------
     alpha : number or tuple of number or list of number or imgaug.parameters.StochasticParameter, optional
-        Visibility of the sharpened image. At 0, only the original image is
-        visible, at 1.0 only its sharpened version is visible.
+        Blending factor of the edge image. At ``0.0``, only the original
+        image is visible, at ``1.0`` only the edge image is visible.
 
-            * If an int or float, exactly that value will be used.
-            * If a tuple ``(a, b)``, a random value from the range ``a <= x <= b`` will
-              be sampled per image.
-            * If a list, then a random value will be sampled from that list
+            * If a number, exactly that value will always be used.
+            * If a tuple ``(a, b)``, a random value will be sampled from the
+              interval ``[a, b]`` per image.
+            * If a list, a random value will be sampled from that list
               per image.
-            * If a StochasticParameter, a value will be sampled from the
+            * If a ``StochasticParameter``, a value will be sampled from that
               parameter per image.
 
     direction : number or tuple of number or list of number or imgaug.parameters.StochasticParameter, optional
-        Angle of edges to pronounce, where 0 represents 0 degrees and 1.0
-        represents 360 degrees (both clockwise, starting at the top).
-        Default value is ``(0.0, 1.0)``, i.e. pick a random angle per image.
+        Angle (in degrees) of edges to pronounce, where ``0`` represents
+        ``0`` degrees and ``1.0`` represents 360 degrees (both clockwise,
+        starting at the top). Default value is ``(0.0, 1.0)``, i.e. pick a
+        random angle per image.
 
-            * If an int or float, exactly that value will be used.
-            * If a tuple ``(a, b)``, a random value from the range ``a <= x <= b`` will
-              be sampled per image.
+            * If a number, exactly that value will always be used.
+            * If a tuple ``(a, b)``, a random value will be sampled from the
+              interval ``[a, b]`` will be sampled per image.
             * If a list, then a random value will be sampled from that list
               per image.
-            * If a StochasticParameter, a value will be sampled from the
+            * If a ``StochasticParameter``, a value will be sampled from the
               parameter per image.
 
     name : None or str, optional
@@ -495,39 +570,43 @@ def DirectedEdgeDetect(alpha=0, direction=(0.0, 1.0), name=None, deterministic=F
 
     Examples
     --------
-    >>> aug = DirectedEdgeDetect(alpha=1.0, direction=0)
+    >>> import imgaug.augmenters as iaa
+    >>> aug = iaa.DirectedEdgeDetect(alpha=1.0, direction=0)
 
-    turns input images into edge images in which edges are detected from
-    top side of the image (i.e. the top sides of horizontal edges are
-    added to the output).
+    Turn input images into edge images in which edges are detected from
+    the top side of the image (i.e. the top sides of horizontal edges are
+    part of the edge image, while vertical edges are ignored).
 
-    >>> aug = DirectedEdgeDetect(alpha=1.0, direction=90/360)
+    >>> aug = iaa.DirectedEdgeDetect(alpha=1.0, direction=90/360)
 
-    same as before, but detecting edges from the right (right side of each
-    vertical edge).
+    Same as before, but edges are detected from the right. Horizontal edges
+    are now ignored.
 
-    >>> aug = DirectedEdgeDetect(alpha=1.0, direction=(0.0, 1.0))
+    >>> aug = iaa.DirectedEdgeDetect(alpha=1.0, direction=(0.0, 1.0))
 
-    same as before, but detecting edges from a variable direction (anything
-    between 0 and 1.0, i.e. 0 degrees and 360 degrees, starting from the
-    top and moving clockwise).
+    Same as before, but edges are detected from a random angle sampled
+    uniformly from the interval ``[0deg, 360deg]``.
 
-    >>> aug = DirectedEdgeDetect(alpha=(0.0, 0.3), direction=0)
+    >>> aug = iaa.DirectedEdgeDetect(alpha=(0.0, 0.3), direction=0)
 
-    generates edge images (edges detected from the top) and overlays them
-    with the input images by a variable amount between 0 and 30 percent
-    (e.g. for 0.3 then ``0.7*old_image + 0.3*edge_image``).
+    Similar to the previous examples, but here the edge image is alpha-blended
+    with the input image. The result is a mixture between the edge image and
+    the input image. The blending factor is randomly sampled between ``0%``
+    and ``30%``.
 
     """
-    alpha_param = iap.handle_continuous_param(alpha, "alpha", value_range=(0, 1.0), tuple_to_uniform=True,
-                                              list_to_choice=True)
-    direction_param = iap.handle_continuous_param(direction, "direction", value_range=None, tuple_to_uniform=True,
-                                                  list_to_choice=True)
+    alpha_param = iap.handle_continuous_param(
+        alpha, "alpha",
+        value_range=(0, 1.0), tuple_to_uniform=True, list_to_choice=True)
+    direction_param = iap.handle_continuous_param(
+        direction, "direction",
+        value_range=None, tuple_to_uniform=True, list_to_choice=True)
 
     def create_matrices(_image, nb_channels, random_state_func):
         alpha_sample = alpha_param.draw_sample(random_state=random_state_func)
         ia.do_assert(0 <= alpha_sample <= 1.0)
-        direction_sample = direction_param.draw_sample(random_state=random_state_func)
+        direction_sample = direction_param.draw_sample(
+            random_state=random_state_func)
 
         deg = int(direction_sample * 360) % 360
         rad = np.deg2rad(deg)
@@ -540,14 +619,14 @@ def DirectedEdgeDetect(alpha=0, direction=(0.0, 1.0), name=None, deterministic=F
             [0, 0, 0],
             [0, 0, 0]
         ], dtype=np.float32)
-        for x in [-1, 0, 1]:
-            for y in [-1, 0, 1]:
-                if (x, y) != (0, 0):
-                    cell_vector = np.array([x, y])
-                    distance_deg = np.rad2deg(ia.angle_between_vectors(cell_vector, direction_vector))
-                    distance = distance_deg / 180
-                    similarity = (1 - distance)**4
-                    matrix_effect[y+1, x+1] = similarity
+        for x, y in itertools.product([-1, 0, 1], [-1, 0, 1]):
+            if (x, y) != (0, 0):
+                cell_vector = np.array([x, y])
+                distance_deg = np.rad2deg(
+                    ia.angle_between_vectors(cell_vector, direction_vector))
+                distance = distance_deg / 180
+                similarity = (1 - distance)**4
+                matrix_effect[y+1, x+1] = similarity
         matrix_effect = matrix_effect / np.sum(matrix_effect)
         matrix_effect = matrix_effect * (-1)
         matrix_effect[1, 1] = 1
@@ -558,11 +637,18 @@ def DirectedEdgeDetect(alpha=0, direction=(0.0, 1.0), name=None, deterministic=F
             [0, 0, 0]
         ], dtype=np.float32)
 
-        matrix = (1-alpha_sample) * matrix_nochange + alpha_sample * matrix_effect
+        matrix = (
+            (1-alpha_sample) * matrix_nochange
+            + alpha_sample * matrix_effect
+        )
 
         return [matrix] * nb_channels
 
     if name is None:
         name = "Unnamed%s" % (ia.caller_name(),)
 
-    return Convolve(create_matrices, name=name, deterministic=deterministic, random_state=random_state)
+    return Convolve(
+        create_matrices,
+        name=name,
+        deterministic=deterministic,
+        random_state=random_state)